+++ CHANGES.md
@@ -16,6 +16,7 @@ Fixed:
 - Fix audio/video synchronization issues when decoding live streams using ffmpeg.
 - Fix issues with TLS connecting clients not being properly timed out (#3598)
 - Make sure reconnection errors are router through the regulat `on_error` callback in `output.icecast` (#3635)
+- Fixed discontinuity count after a restart in HLS outputs.
 
 ---